DRILL-3051: Fix integer overflow in TimedRunnable
also fixes a similar overflow in Foreman

+++ exec/java-exec/src/main/java/org/apache/drill/exec/store/TimedRunnable.java
@@ -41,7 +41,7 @@ import com.google.common.collect.Lists;
  */
 public abstract class TimedRunnable<V> implements Runnable {
 
-  private static int TIMEOUT_PER_RUNNABLE_IN_MSECS = 15000;
+  private static long TIMEOUT_PER_RUNNABLE_IN_MSECS = 15000;
 
   private volatile Exception e;
   private volatile long timeNanos;
